log_user_event take enum parameter not i32

+++ src/api/core/two_factor/duo.rs
@@ -128,7 +128,7 @@ async fn activate_duo(data: Json<EnableDuoData>, headers: Headers, conn: DbConn)
 
     generate_recover_code(&mut user, &conn).await;
 
-    log_user_event(EventType::UserUpdated2fa as i32, &user.uuid, headers.device.atype, &headers.ip.ip, &conn).await;
+    log_user_event(EventType::UserUpdated2fa, &user.uuid, headers.device.atype, &headers.ip.ip, &conn).await;
 
     Ok(Json(json!({
         "duo": json!({
@@ -160,8 +160,7 @@ async fn disable_duo(data: Json<VerificationTokenData>, headers: Headers, conn: 
         two_factor::validate_duo(&data.user_verification_token, &user.uuid, duo.as_ref(), true)?;
 
         twofactor.delete(&conn).await?;
-        log_user_event(EventType::UserDisabled2fa as i32, &user.uuid, headers.device.atype, &headers.ip.ip, &conn)
-            .await;
+        log_user_event(EventType::UserDisabled2fa, &user.uuid, headers.device.atype, &headers.ip.ip, &conn).await;
     }
 
     if TwoFactor::find_by_user(&user.uuid, &conn).await.is_empty() {
